+++ idea/testData/debugger/tinyApp/src/evaluate/singleBreakpoint/javaContext/jcImports.kt
@@ -1,0 +1,13 @@
+package jcImports
+
+fun main(args: Array<String>) {
+    val javaClass = forTests.javaContext.JavaClass()
+    //Breakpoint!
+    javaClass.imports()
+}
+
+// STEP_INTO: 1
+// STEP_OVER: 1
+
+// EXPRESSION: list.filter { it == 1 }.size
+// RESULT: 1: I